+++ CHANGELOG.md
@@ -46,6 +46,9 @@ Group changes to describe their impact on the project, as follows:
 - **Voicemail server:** New server for recording voice messages intended for the recipient of the call when a call is 
   declined or not answered. See [Voicemail Server](https://wiki.linphone.org/xwiki/wiki/public/view/Flexisip/C.%20Features/Voicemail%20server/)
   for more information.
+- **B2BUA:** Add a Priority header to NOTIFY messages containing Message Waiting Indication (MWI) information. If the
+  MWI notifies new urgent messages, the Priority header is set to 'urgent', otherwise it is set to 'non-urgent'. This
+  has the effect of only sending push notifications for MWI indicating urgent messages.
 - **Build:** Add support for C++20
 - **CLI:** Add three new commands to manage entries in the passwords cache.
     - `AUTH_CACHE_LIST`: list all entries in the cache (optional: filter by domain)